feat: 이미지 업로드 정책 상수 분리 및 타입명 정리
